CHORE: Factor out golden test macros
