+++ Cargo.toml
@@ -14,6 +14,8 @@ license-file = "LICENSE-APACHE"
 anyhow = "1"
 async-trait = "0.1"
 async-stm = "0.2"
+arbitrary = { version = "1", features = ["derive"] }
+arbtest = "0.2"
 futures = "0.3"
 paste = "1"
 serde = { version = "1", features = ["derive"] }
@@ -23,7 +25,6 @@ tempfile = "3.3"
 thiserror = "1"
 quickcheck = "1"
 quickcheck_macros = "1"
-arbitrary = { version = "1", features = ["derive"] }
 libsecp256k1 = "0.7"
 rand = "0.8"
 rand_chacha = "0.3"
