+++ fendermint/testing/Cargo.toml
@@ -9,7 +9,13 @@ license.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-arbitrary = { version = "1.2" }
+serde = { workspace = true }
+arbitrary = { workspace = true }
+quickcheck = { workspace = true }
+hex = { workspace = true }
+
+cid = { workspace = true }
+fvm_ipld_encoding = { workspace = true }
 
 [dev-dependencies]
-arbtest = { version = "0.2" }
+arbtest = { workspace = true }
