+++ fendermint/testing/src/lib.rs
@@ -1,3 +1,4 @@
 // Copyright 2022-2023 Protocol Labs
 // SPDX-License-Identifier: Apache-2.0, MIT
+pub mod golden;
 pub mod smt;
